Rearrange Methods
Just a thought: It might make more sense to order the methods in the flow and lifespan of content that is submarined. I simply moved uploadFileOrFolder and uploadJson to the top of the methods. This makes sense because you can't query any content if you haven't uploaded it yet. Now the methods are in the flow of how someone would use submarining: upload, getContent, generateLink, make small changes, delete